refactor: restructures app organization
